fix: load the LLM eval fixtures without pytest_plugins so the whole tests tree collects
